+++ benchmarks/consumer-bundles/v1/root-plus-backbone-jquery.js
@@ -1,5 +1,5 @@
 import * as Marionette from 'marionette';
 
-export { default as Backbone } from 'marionette/backbone';
-export { default as jQueryDomApi } from 'marionette/jquery-dom-api';
+export { default as BackboneApi } from '@marionette/adapters/backbone';
+export { default as jQueryDomApi } from '@marionette/adapters/dom/jquery';
 export { Marionette };